generic room api for space relations
