+++ native/acter/src/api/convo.rs
@@ -43,7 +43,7 @@ use std::{
 };
 use tracing::{error, info, trace, warn};
 
-use crate::TimelineStream;
+use crate::{SpaceRelations, TimelineStream};
 
 use super::{
     client::Client,
