+++ native/acter/src/api/room.rs
@@ -1,4 +1,9 @@
+pub use acter_core::spaces::{
+    CreateSpaceSettings, CreateSpaceSettingsBuilder, RelationTargetType, SpaceRelation,
+    SpaceRelations as CoreSpaceRelations,
+};
 use acter_core::{
+    client::CoreClient,
     events::{
         calendar::CalendarEventEventContent,
         news::{NewsContent, NewsEntryEvent, NewsEntryEventContent},
@@ -17,9 +22,19 @@ use matrix_sdk::{
     media::{MediaFormat, MediaRequest},
     room::{Room as SdkRoom, RoomMember},
     ruma::{
+        api::client::space::{
+            get_hierarchy::v1::{Request, Response},
+            SpaceHierarchyRoomsChunk, SpaceRoomJoinRule,
+        },
         api::client::{
             receipt::create_receipt::v3::ReceiptType as CreateReceiptType,
             room::report_content::v3::Request as ReportContentRequest,
+        },
+        api::client::{
+            space::get_hierarchy::v1::{
+                Request as GetHierarchyRequest, Response as GetHierarchyResponse,
+            },
+            state::send_state_event::v3::Request as SendStateEventRequest,
         },
         assign,
         events::{
@@ -40,13 +55,18 @@ use matrix_sdk::{
             AnyMessageLikeEvent, AnyStateEvent, AnyTimelineEvent, MessageLikeEvent,
             MessageLikeEventType, StateEvent, StateEventType, StaticEventContent,
         },
+        events::{room::MediaSource, space::child::HierarchySpaceChildEvent},
         room::RoomType,
-        EventId, Int, OwnedEventId, OwnedMxcUri, OwnedUserId, TransactionId, UInt, UserId,
+        serde::Raw,
+        EventId, Int, OwnedEventId, OwnedMxcUri, OwnedRoomAliasId, OwnedRoomId, OwnedUserId,
+        TransactionId, UInt, UserId,
     },
     Client, RoomMemberships, RoomState,
 };
 use std::{io::Write, ops::Deref, path::PathBuf};
 use tracing::{error, info};
+
+use crate::OptionBuffer;
 
 use super::{
     account::Account,
@@ -259,18 +279,206 @@ impl Member {
     }
 }
 
+pub struct SpaceHierarchyRoomInfo {
+    chunk: SpaceHierarchyRoomsChunk,
+    client: CoreClient,
+}
+
+impl SpaceHierarchyRoomInfo {
+    pub fn canonical_alias(&self) -> Option<OwnedRoomAliasId> {
+        self.chunk.canonical_alias.clone()
+    }
+
+    /// The name of the room, if any.
+    pub fn name(&self) -> Option<String> {
+        self.chunk.name.clone()
+    }
+
+    /// The number of members joined to the room.
+    pub fn num_joined_members(&self) -> u64 {
+        self.chunk.num_joined_members.into()
+    }
+
+    /// The ID of the room.
+    pub fn room_id(&self) -> OwnedRoomId {
+        self.chunk.room_id.clone()
+    }
+
+    pub fn room_id_str(&self) -> String {
+        self.room_id().to_string()
+    }
+
+    pub fn topic(&self) -> Option<String> {
+        self.chunk.topic.clone()
+    }
+
+    /// Whether the room may be viewed by guest users without joining.
+    pub fn world_readable(&self) -> bool {
+        self.chunk.world_readable
+    }
+
+    pub fn guest_can_join(&self) -> bool {
+        self.chunk.guest_can_join
+    }
+
+    pub fn avatar_url(&self) -> Option<OwnedMxcUri> {
+        self.chunk.avatar_url.clone()
+    }
+
+    pub fn avatar_url_str(&self) -> Option<String> {
+        self.avatar_url().map(|a| a.to_string())
+    }
+
+    /// The join rule of the room.
+    pub fn join_rule(&self) -> SpaceRoomJoinRule {
+        self.chunk.join_rule.clone()
+    }
+
+    pub fn join_rule_str(&self) -> String {
+        self.join_rule().to_string()
+    }
+
+    /// The type of room from `m.room.create`, if any.
+    pub fn room_type(&self) -> Option<RoomType> {
+        self.chunk.room_type.clone()
+    }
+
+    pub fn is_space(&self) -> bool {
+        matches!(self.chunk.room_type, Some(RoomType::Space))
+    }
+
+    /// The stripped `m.space.child` events of the space-room.
+    ///
+    /// If the room is not a space-room, this should be empty.
+    pub fn children_state(&self) -> Vec<Raw<HierarchySpaceChildEvent>> {
+        self.chunk.children_state.clone()
+    }
+
+    pub fn has_avatar(&self) -> bool {
+        self.chunk.avatar_url.is_some()
+    }
+
+    pub fn via_server_name(&self) -> Option<String> {
+        for v in &self.chunk.children_state {
+            let Ok(h) = v.deserialize() else { continue };
+            let Some(via) = h.content.via else { continue };
+            if let Some(v) = via.into_iter().next() {
+                return Some(v.to_string());
+            }
+        }
+        None
+    }
+
+    pub async fn get_avatar(&self) -> Result<OptionBuffer> {
+        let client = self.client.client().clone();
+        if let Some(url) = self.chunk.avatar_url.clone() {
+            return RUNTIME
+                .spawn(async move {
+                    let request = MediaRequest {
+                        source: MediaSource::Plain(url),
+                        format: MediaFormat::File,
+                    };
+                    let buf = client.media().get_media_content(&request, true).await?;
+                    Ok(OptionBuffer::new(Some(buf)))
+                })
+                .await?;
+        }
+        Ok(OptionBuffer::new(None))
+    }
+}
+
+impl SpaceHierarchyRoomInfo {
+    pub(crate) async fn new(chunk: SpaceHierarchyRoomsChunk, client: CoreClient) -> Self {
+        SpaceHierarchyRoomInfo { chunk, client }
+    }
+}
+
+pub struct SpaceHierarchyListResult {
+    resp: GetHierarchyResponse,
+    client: CoreClient,
+}
+
+impl SpaceHierarchyListResult {
+    pub fn next_batch(&self) -> Option<String> {
+        self.resp.next_batch.clone()
+    }
+
+    pub async fn rooms(&self) -> Result<Vec<SpaceHierarchyRoomInfo>> {
+        let client = self.client.clone();
+        let chunks = self.resp.rooms.clone();
+        RUNTIME
+            .spawn(async move {
+                let iter = chunks
+                    .into_iter()
+                    .map(|chunk| SpaceHierarchyRoomInfo::new(chunk, client.clone()));
+                Ok(futures::future::join_all(iter).await)
+            })
+            .await?
+    }
+}
+
+pub struct SpaceRelations {
+    pub(crate) core: CoreSpaceRelations,
+    pub(crate) room: Room,
+}
+
+impl Deref for SpaceRelations {
+    type Target = CoreSpaceRelations;
+    fn deref(&self) -> &Self::Target {
+        &self.core
+    }
+}
+
+impl SpaceRelations {
+    pub fn room_id(&self) -> OwnedRoomId {
+        self.room.room_id().to_owned()
+    }
+
+    pub fn room_id_str(&self) -> String {
+        self.room.room_id().to_string()
+    }
+
+    pub async fn query_hierarchy(&self, from: Option<String>) -> Result<SpaceHierarchyListResult> {
+        let c = self.room.core.clone();
+        let room_id = self.room.room_id().to_owned();
+        RUNTIME
+            .spawn(async move {
+                let request = assign!(GetHierarchyRequest::new(room_id), { from, max_depth: Some(1u32.into()) });
+                let resp = c.client().send(request, None).await?;
+                Ok(SpaceHierarchyListResult { resp, client: c.clone() })
+            })
+            .await?
+    }
+}
+
 #[derive(Clone, Debug)]
 pub struct Room {
+    pub(crate) core: CoreClient,
     pub(crate) room: SdkRoom,
 }
 
 impl Room {
+    pub fn new(core: CoreClient, room: SdkRoom) -> Self {
+        Room { core, room }
+    }
+
     pub async fn is_acter_space(&self) -> Result<bool> {
         let inner = self.room.clone();
         let result = RUNTIME
             .spawn(async move { is_acter_space(&inner).await })
             .await?;
         Ok(result)
+    }
+
+    pub async fn space_relations(&self) -> Result<SpaceRelations> {
+        let c = self.core.clone();
+        let me = self.clone();
+        RUNTIME
+            .spawn(async move {
+                let core = c.space_relations(&me.room).await?;
+                Ok(SpaceRelations { core, room: me })
+            })
+            .await?
     }
 
     pub async fn get_my_membership(&self) -> Result<Member> {
