+++ native/acter/src/api/spaces.rs
@@ -446,185 +446,12 @@ impl Space {
     }
 }
 
-pub struct SpaceRelations {
-    core: CoreSpaceRelations,
-    space: Space,
-}
-
-impl Deref for SpaceRelations {
-    type Target = CoreSpaceRelations;
-    fn deref(&self) -> &Self::Target {
-        &self.core
-    }
-}
-
-pub struct SpaceHierarchyRoomInfo {
-    chunk: SpaceHierarchyRoomsChunk,
-    client: Client,
-}
-
-impl SpaceHierarchyRoomInfo {
-    pub fn canonical_alias(&self) -> Option<OwnedRoomAliasId> {
-        self.chunk.canonical_alias.clone()
-    }
-
-    /// The name of the room, if any.
-    pub fn name(&self) -> Option<String> {
-        self.chunk.name.clone()
-    }
-
-    /// The number of members joined to the room.
-    pub fn num_joined_members(&self) -> u64 {
-        self.chunk.num_joined_members.into()
-    }
-
-    /// The ID of the room.
-    pub fn room_id(&self) -> OwnedRoomId {
-        self.chunk.room_id.clone()
-    }
-
-    pub fn room_id_str(&self) -> String {
-        self.room_id().to_string()
-    }
-
-    pub fn topic(&self) -> Option<String> {
-        self.chunk.topic.clone()
-    }
-
-    /// Whether the room may be viewed by guest users without joining.
-    pub fn world_readable(&self) -> bool {
-        self.chunk.world_readable
-    }
-
-    pub fn guest_can_join(&self) -> bool {
-        self.chunk.guest_can_join
-    }
-
-    pub fn avatar_url(&self) -> Option<OwnedMxcUri> {
-        self.chunk.avatar_url.clone()
-    }
-
-    pub fn avatar_url_str(&self) -> Option<String> {
-        self.avatar_url().map(|a| a.to_string())
-    }
-
-    /// The join rule of the room.
-    pub fn join_rule(&self) -> SpaceRoomJoinRule {
-        self.chunk.join_rule.clone()
-    }
-
-    pub fn join_rule_str(&self) -> String {
-        self.join_rule().to_string()
-    }
-
-    /// The type of room from `m.room.create`, if any.
-    pub fn room_type(&self) -> Option<RoomType> {
-        self.chunk.room_type.clone()
-    }
-
-    pub fn is_space(&self) -> bool {
-        matches!(self.chunk.room_type, Some(RoomType::Space))
-    }
-
-    /// The stripped `m.space.child` events of the space-room.
-    ///
-    /// If the room is not a space-room, this should be empty.
-    pub fn children_state(&self) -> Vec<Raw<HierarchySpaceChildEvent>> {
-        self.chunk.children_state.clone()
-    }
-
-    pub fn has_avatar(&self) -> bool {
-        self.chunk.avatar_url.is_some()
-    }
-
-    pub fn via_server_name(&self) -> Option<String> {
-        for v in &self.chunk.children_state {
-            let Ok(h) = v.deserialize() else { continue };
-            let Some(via) = h.content.via else { continue };
-            if let Some(v) = via.into_iter().next() {
-                return Some(v.to_string());
-            }
-        }
-        None
-    }
-
-    pub async fn get_avatar(&self) -> Result<OptionBuffer> {
-        let client = self.client.clone();
-        if let Some(url) = self.chunk.avatar_url.clone() {
-            return RUNTIME
-                .spawn(async move {
-                    let request = MediaRequest {
-                        source: MediaSource::Plain(url),
-                        format: MediaFormat::File,
-                    };
-                    let buf = client.media().get_media_content(&request, true).await?;
-                    Ok(OptionBuffer::new(Some(buf)))
-                })
-                .await?;
-        }
-        Ok(OptionBuffer::new(None))
-    }
-}
-
-impl SpaceHierarchyRoomInfo {
-    pub(crate) async fn new(chunk: SpaceHierarchyRoomsChunk, client: Client) -> Self {
-        SpaceHierarchyRoomInfo { chunk, client }
-    }
-}
-
-pub struct SpaceHierarchyListResult {
-    resp: GetHierarchyResponse,
-    client: Client,
-}
-
-impl SpaceHierarchyListResult {
-    pub fn next_batch(&self) -> Option<String> {
-        self.resp.next_batch.clone()
-    }
-
-    pub async fn rooms(&self) -> Result<Vec<SpaceHierarchyRoomInfo>> {
-        let client = self.client.clone();
-        let chunks = self.resp.rooms.clone();
-        RUNTIME
-            .spawn(async move {
-                let iter = chunks
-                    .into_iter()
-                    .map(|chunk| SpaceHierarchyRoomInfo::new(chunk, client.clone()));
-                Ok(futures::future::join_all(iter).await)
-            })
-            .await?
-    }
-}
-
-impl SpaceRelations {
-    pub fn room_id(&self) -> OwnedRoomId {
-        self.space.room_id().to_owned()
-    }
-
-    pub fn room_id_str(&self) -> String {
-        self.space.room_id().to_string()
-    }
-
-    pub async fn query_hierarchy(&self, from: Option<String>) -> Result<SpaceHierarchyListResult> {
-        let c = self.space.client.clone();
-        let room_id = self.space.room_id().to_owned();
-        RUNTIME
-            .spawn(async move {
-                let request = assign!(GetHierarchyRequest::new(room_id), { from, max_depth: Some(1u32.into()) });
-                let resp = c.send(request, None).await?;
-                Ok(SpaceHierarchyListResult { resp, client: c.clone() })
-            })
-            .await?
-    }
-}
-
 // External API
 
 impl Space {
     pub fn new(client: Client, inner: Room) -> Self {
         Space { client, inner }
     }
-
     pub async fn create_onboarding_data(&self) -> Result<()> {
         let mut engine = Engine::with_template(std::include_str!("../templates/onboarding.toml"))?;
         engine
@@ -737,17 +564,6 @@ impl Space {
                     )
                     .await?;
                 Ok(response.event_id.to_string())
-            })
-            .await?
-    }
-
-    pub async fn space_relations(&self) -> Result<SpaceRelations> {
-        let c = self.client.core.clone();
-        let me = self.clone();
-        RUNTIME
-            .spawn(async move {
-                let core = c.space_relations(&me.room).await?;
-                Ok(SpaceRelations { core, space: me })
             })
             .await?
     }
